Combine client and device endpoints (tests)
